Issue #69 - Minor code refactoring
